// src/app/models/giohangsanpham.ts
export class GioHangSanPham{
    maGH?: number;
    idUser?: number;
    maSP?: string;
    soLuong?: number;
    maSize?: number;
    tongTien?: number;
    tenSP?: string;
    kieuDang?: string;
    chatLieu?: string;
    image_URL?: string;
    soLuongKho?: number;
    trangThai?: string;
    tongTienKho?: number;
    soLuongDapUng?: number;
}